refactor: move http event handler to a separate file
